requirement.txt
torch 1.12.1+cu113
torchvision 0.13.1+cu113
matplotlib 3.5.2
Pillow 9.5.0
scikit-learn 1.2.2
numpy 1.23.5